perception/requirements.txt
opencv-python
numpy
torch
torchsummary
pillow==8.2.0
torchvision